split outlier detection into separate steps
