+++ jwst/outlier_detection/__init__.py
@@ -1,3 +1,11 @@
-from .outlier_detection_step import OutlierDetectionStep
+from .outlier_detection_coron_step import OutlierDetectionCoronStep
+from .outlier_detection_ifu_step import OutlierDetectionIFUStep
+from .outlier_detection_spec_step import OutlierDetectionSpecStep
+from .outlier_detection_imaging_step import OutlierDetectionImagingStep
+from .outlier_detection_tso_step import OutlierDetectionTSOStep
 
-__all__ = ['OutlierDetectionStep']
+__all__ = ['OutlierDetectionCoronStep',
+           'OutlierDetectionIFUStep',
+           'OutlierDetectionSpecStep',
+           'OutlierDetectionImagingStep',
+           'OutlierDetectionTSOStep']
